requirements.txt
exceptiongroup==1.2.2
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.4
python-dotenv==1.0.1
tomli==2.2.1